libft without gnl
