+++ libft/.gitignore
@@ -1,10 +1,26 @@
 .DS_Store
-/test
-.normignore
-
-#VSCode
-/.vscode
-/*.dSYM
-
-libft.a
+*.d
 *.o
+*.a
+*.so
+*.dylib
+*.dll
+*.lib
+*.exp
+*.def
+*.manifest
+*.rc
+*.res
+*.rc
+*.res
+*.obj
+*.exe
+*.out
+*.bin
+*.hex
+*.elf
+*.map
+*.os
+*.dt
+*.tar*
+build/**
