+++ libft/src/list_read.c
@@ -1,25 +1,42 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   ft_libs.h                                          :+:      :+:    :+:   */
+/*   list_read.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: ynishimu <ynishimu@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2023/04/12 12:38:20 by ynishimu          #+#    #+#             */
-/*   Updated: 2023/04/12 19:08:57 by ynishimu         ###   ########.fr       */
+/*   Created: 2023/01/16 22:30:49 by ynishimu          #+#    #+#             */
+/*   Updated: 2023/04/12 18:09:35 by ynishimu         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#ifndef FT_LIBS_H
-# define FT_LIBS_H
+#include "libft.h"
 
-# include <stdlib.h>
-# include <unistd.h>
-# include <stddef.h>
-# include <limits.h>
-# include <errno.h>
-# include <stdbool.h>
-# include <stdarg.h>
-# include <stdint.h>
+t_list	*ft_lstlast(t_list *lst)
+{
+	t_list	*node;
 
-#endif
+	if (!lst)
+		return (NULL);
+	node = lst;
+	while (node->next)
+		node = node->next;
+	return (node);
+}
+
+int	ft_lstsize(t_list *lst)
+{
+	t_list	*node;
+	size_t	size;
+
+	if (!lst)
+		return (0);
+	size = 0;
+	node = lst;
+	while (node)
+	{
+		node = node->next;
+		size++;
+	}
+	return (size);
+}
