+++ libft/src/print_number.c
@@ -10,7 +10,7 @@
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "../../includes/libft.h"
+#include "libft.h"
 
 ssize_t	ft_putnbr_fd(int n, int fd)
 {
