+++ libft/src/print_string.c
@@ -1,36 +1,43 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   mem_allocate.c                                     :+:      :+:    :+:   */
+/*   print_string.c                                     :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: ynishimu <ynishimu@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2023/02/25 21:09:42 by ynishimu          #+#    #+#             */
+/*   Created: 2023/01/16 13:23:42 by ynishimu          #+#    #+#             */
 /*   Updated: 2023/04/12 18:09:35 by ynishimu         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "../../includes/libft.h"
+#include "libft.h"
 
-void	*ft_malloc(size_t size, void *asnd_0, void *asnd_1, void *asnd_2)
+ssize_t	ft_putchar_fd(char c, int fd)
 {
-	void	*allocated;
-
-	allocated = malloc(size);
-	if (!allocated)
-		ft_exit(-1, asnd_0, asnd_1, asnd_2);
-	return (allocated);
+	if (fd < 0)
+		return (0);
+	return (write(fd, &c, 1));
 }
 
-void	*ft_calloc(size_t count, size_t size)
+ssize_t	ft_putstr_fd(char *s, int fd)
 {
-	void	*alcd_ptr;
+	if (!s || fd < 0)
+		return (0);
+	return (write(fd, s, ft_strlen(s)));
+}
 
-	if (size != 0 && count > SIZE_MAX / size)
-		return (NULL);
-	alcd_ptr = malloc(count * size);
-	if (!alcd_ptr)
-		return (NULL);
-	ft_bzero(alcd_ptr, count * size);
-	return (alcd_ptr);
+ssize_t	ft_putendl_fd(char *s, int fd)
+{
+	ssize_t	printed;
+
+	printed = 0;
+	if (!s || fd < 0)
+		return (-1);
+	printed += ft_putstr_fd(s, fd);
+	if (printed < 0)
+		return (printed);
+	printed += ft_putchar_fd('\n', fd);
+	if (printed < 0)
+		return (printed);
+	return (printed);
 }
